Implemented some fault-testing for normal surfaces
Implement `separates`, `isEssentialSphere`, `isEssentialTorus`, and
`isSolidTorusAnnulus` member functions of the `NormalSurface` class.

`separates` determines whether or not cutting along the surface adds a
new component. This is not the same as testing whether or not the
exterior of the surface is disconnected.

`isEssentialSphere` was necessary in addition to the member function
`hasNonTrivialSphereOrDisc` of the `Triangulation<3>` class. Essential
spheres are nontrivial, but not the other way round.

`isEssentialTorus` can't be implemented with just a call to
`isIncompressible`, since boundary-parallel tori are incompressible
yet not essential. The `isTorusXInterval` member function of
`Triangulation<3>` makes `isEssentialTorus` possible.

`isSolidTorusAnnulus` does not determine whether or not an annulus is
essential, but only whether or not a surface is an annulus cutting the
ambient manifold into one or two solid tori. Determining whether or
not an annulus is essential would require boundary patterns. Though
that is a worthy goal, it is not necessary for hyperbolicity
testing. It suffices to implement the auxiliary function
`isSolidTorusAnnulus`.

In the prime and JSJ decompositions one cuts along essential surfaces
of nonnegative Euler characteristic. Moreover, Weeks's Newton's-method
applied to triangulations of such manifolds tends to ``pull apart''
those tetrahedra containing quads of such surfaces. In materials science
and geology, the region where a solid (or the earth) separates is a
"fault." So an essential surface of nonnegative Euler characteristic I
also call a "fault."